Gitignore transpiled unit tests, rename test spec directory
