+++ .gitignore
@@ -1,3 +1,4 @@
 lib/
 node_modules/
+test/spec-build/
 npm-debug.log
