+++ karma.conf.js
@@ -11,7 +11,7 @@ module.exports = function(config) {
 
     // list of files / patterns to load in the browser
     files: [
-      'test/spec/*.spec.js',
+      'test/spec-build/*.spec.js',
       { pattern : 'test/thread-scripts/*.js', included : false }
     ],
 
@@ -23,7 +23,7 @@ module.exports = function(config) {
     // preprocess matching files before serving them to the browser
     // available preprocessors: https://npmjs.org/browse/keyword/karma-preprocessor
     preprocessors: {
-      'test/spec/*.spec.js': ['browserify']
+      'test/spec-build/*.spec.js': ['browserify']
     },
 
     browserify: {
