+++ test/spec/pool-functional.spec.js
@@ -1,27 +1,19 @@
-'use strict';
+import expect   from 'expect.js';
+import { Pool } from '../../lib/';
 
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
+describe('Pool (functional test)', () => {
+  const pool = new Pool();
+  const jobs = [], promises = [];
 
-var _expectJs = require('expect.js');
-
-var _expectJs2 = _interopRequireDefault(_expectJs);
-
-var _lib = require('../../lib/');
-
-describe('Pool (functional test)', function () {
-  var pool = new _lib.Pool();
-  var jobs = [],
-      promises = [];
-
-  var handler = function handler(input, done) {
+  const handler = (input, done) => {
     done(input);
   };
 
   pool.run(handler);
 
-  it('can send data and run promisified', function () {
-    for (var i = 0; i < 10; i++) {
-      var job = pool.send(i);
+  it('can send data and run promisified', () => {
+    for (let i = 0; i < 10; i++) {
+      const job = pool.send(i);
       if (jobs.indexOf(job) > -1) {
         throw new Error('pool.send() returns the same job twice');
       }
@@ -31,12 +23,15 @@ describe('Pool (functional test)', function () {
     }
   });
 
-  it('responds as expected', function (done) {
-    Promise.all(promises).then(function (responses) {
-      (0, _expectJs2['default'])(responses.sort()).to.eql([0, 1, 2, 3, 4, 5, 6, 7, 8, 9]);
-      done();
-    })['catch'](function (error) {
-      done(error);
-    });
+  it('responds as expected', done => {
+    Promise
+      .all(promises)
+      .then(responses => {
+        expect(responses.sort()).to.eql([ 0, 1, 2, 3, 4, 5, 6, 7, 8, 9 ]);
+        done();
+      })
+      .catch(error => {
+        done(error);
+      });
   });
 });
